backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Issam"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
google-auth = "^2.23.3"
google-cloud = "^0.34.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"